new change and add Subject lecure model and subject presence

+++ src/main/java/vp/spring/rcs/TestData.java
@@ -87,7 +87,7 @@ public class TestData {
 		interstellarSpace.getStyles().add(avantGardeJazz);
 		recordService.save(interstellarSpace);
 		
-		SecurityUser user1 = new SecurityUser(1, "User", "$2a$04$4pqDFh9SxLAg/uSH59JCB.LwIS6QoAjM9qcE7H9e2drFuWhvTnDFi", "Zoran", "Peric");
+		SecurityUser user1 = new SecurityUser((long)1, "User", "$2a$04$4pqDFh9SxLAg/uSH59JCB.LwIS6QoAjM9qcE7H9e2drFuWhvTnDFi", "Zoran", "Peric");
 		securityUserService.save(user1);
 		
 		SecurityAuthority authority1 = new SecurityAuthority(1, "User");
@@ -96,7 +96,7 @@ public class TestData {
 		SecurityUserAuthority userAuthority1 = new SecurityUserAuthority(1, user1, authority1);
 		securityUserAuthorityService.save(userAuthority1);
 		
-		Student student1 = new Student(1, "pera", "$2a$04$4pqDFh9SxLAg/uSH59JCB.LwIS6QoAjM9qcE7H9e2drFuWhvTnDFi", "Petar", "Petrovic", 123456789, "SF100");
+		Student student1 = new Student((long)1, "pera", "$2a$04$4pqDFh9SxLAg/uSH59JCB.LwIS6QoAjM9qcE7H9e2drFuWhvTnDFi", "Petar", "Petrovic", 123456789, "SF100");
 		studentService.save(student1);
 	}
